# openapi.yaml
openapi: 3.1.0
info:
  title: Ticket API
  version: 0.1.891
  description: |
    RESTful API for the Ticket project and issue management system.
    All endpoints are under `/api/`. Authentication uses Bearer tokens
    or session cookies (`__Host-session` on secure requests; legacy
    `ticket_token` remains accepted).
    Agent endpoints use HTTP Basic Auth (agent UUID + password).
  contact:
    name: Simon Gauld
    url: https://github.com/simonski/ticket
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Session token from /api/login
    CookieAuth:
      type: apiKey
      in: cookie
      name: __Host-session
    BasicAuth:
      type: http
      scheme: basic
      description: Agent UUID as username, agent password as password
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
      example:
        error: "ticket not found"
    User:
      type: object
      example:
        user_id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        username: "alice"
        role: "user"
        display_name: "Alice Smith"
        enabled: true
        user_type: "user"
      properties:
        user_id:
          type: string
          format: uuid
        username:
          type: string
        role:
          type: string
          enum: [admin, user]
        display_name:
          type: string
        enabled:
          type: boolean
        user_type:
          type: string
          enum: [user, agent]
        email:
          type: string
        status:
          type: string
        description:
          type: string
        last_seen:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    Agent:
      allOf:
        - $ref: '#/components/schemas/User'
        - type: object
          properties:
            uuid:
              type: string
              format: uuid
...
    AuthResponse:
      type: object
      required: [token, user]
      properties:
        token:
          type: string
        user:
          $ref: '#/components/schemas/User'
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient permissions for this operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  examples:
...
    LoginRequestExample:
      summary: Login request body
      value:
        username: "alice"
        password: "s3cr3tpassword"
    LoginResponseExample:
      summary: Successful login response
      value:
        token: "dGhpcyBpcyBhIGJhc2U2NHRva2Vu"
        user:
          user_id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
          username: "alice"
          role: "user"
...
paths:
...
  /api/register:
    post:
      tags: [Auth]
      summary: Register a new user account
      operationId: register
      description: Registration must be enabled via system settings.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
      responses:
        '201':
          description: User created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Registration disabled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limited
  /api/login:
    post:
      tags: [Auth]
      summary: Authenticate and receive a session token
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username:
                  type: string
                password:
                  type: string
            examples:
              typical:
                $ref: '#/components/examples/LoginRequestExample'
      responses:
        '200':
          description: Authenticated successfully
          headers:
            Set-Cookie:
              description: "__Host-session session cookie (secure requests); legacy ticket_token also accepted (HttpOnly, SameSite=Lax, 30-day expiry)"
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
              examples:
                success:
                  $ref: '#/components/examples/LoginResponseExample'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limited
...
  /api/agents:
    get:
      tags: [Agents]
      summary: List all agents
      operationId: listAgents
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Agent list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Agents]
      summary: Create a new agent
      operationId: createAgent
      security:
        - BearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  description: Optional; auto-generated if omitted
      responses:
        '201':
          description: Agent created
          content:
            application/json:
              schema:
                type: object
                properties:
                  agent:
                    $ref: '#/components/schemas/Agent'
                  password:
                    type: string

        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/agents/register:
    post:
      tags: [Agents]
      summary: Register an agent (agent auth)
      operationId: registerAgent
      security:
        - BasicAuth: []
      responses:
        '200':
          description: Agent registered
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  agent:
                    $ref: '#/components/schemas/Agent'

        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/ws:
    get:
      tags: [WebSocket]
      summary: Live update WebSocket
      operationId: liveUpdates
      description: |
        WebSocket endpoint for real-time ticket and project change events.
        Authenticate via Bearer token or session cookie.
      parameters:
        - name: token
          in: query
          schema:
            type: string
          description: Optional auth token
      responses:
        '101':
          description: WebSocket upgrade

        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/chat/ws:
    get:
      tags: [WebSocket]
      summary: Chat/AI streaming WebSocket
      operationId: chatWebSocket
      description: |
        WebSocket endpoint for streaming LLM chat sessions.
        Chat must be enabled. Messages are JSON with type, text, stream, and error fields.
      parameters:
        - name: token
          in: query
          schema:
            type: string
          description: Optional auth token
      responses:
        '101':
          description: WebSocket upgrade
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'